requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.12
Flask-Executor==1.0.0
Flask-Migrate==4.1.0
SQLAlchemy==2.0.43
python-dotenv==1.1.1
bleach==6.2.0
itsdangerous==2.2.0
APScheduler==3.11.0
google-cloud-secret-manager==2.25.0
google-cloud-storage==3.9.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
pandas==2.3.2
openpyxl==3.1.5
qrcode[pil]==8.2
xhtml2pdf==0.2.17
pytz==2025.2
